# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lighthouse",
    "description": "Knowledge base for AI agents. Public retrieval over HTTP (MCP wrapper on top), proposal pipeline behind a shared API key.",
    "version": "0.0.1"
  },
  "paths": {
...
    "/v1/propose": {
      "post": {
        "tags": [
          "proposal"
        ],
        "summary": "Propose",
        "operationId": "propose_v1_propose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Proposal"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalReceipt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
    "/admin/importers/types": {
      "get": {
        "tags": [
          "admin",
          "importers"
        ],
        "summary": "List Types",
        "description": "All importer types the registry knows about.",
        "operationId": "list_types_admin_importers_types_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImporterTypeOut"
                  },
                  "title": "Response List Types Admin Importers Types Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/importers/discover": {
      "post": {
        "tags": [
          "admin",
          "importers"
        ],
        "summary": "Discover Route",
        "description": "Probe the source with provided creds and list available items.\n\nNo DB writes \u2014 secrets stay in memory only. The wizard sends them\nagain on save (where they get encrypted).",
        "operationId": "discover_route_admin_importers_discover_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoverIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/importers/": {
      "get": {
        "tags": [
          "admin",
          "importers"
        ],
        "summary": "List Route",
        "operationId": "list_route_admin_importers__get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImporterOut"
                  },
                  "title": "Response List Route Admin Importers  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin",
          "importers"
        ],
        "summary": "Create Route",
        "operationId": "create_route_admin_importers__post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImporterCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImporterOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/importers/{importer_id}": {
      "get": {
        "tags": [
          "admin",
          "importers"
        ],
        "summary": "Get Route",
        "operationId": "get_route_admin_importers__importer_id__get",
        "parameters": [
          {
            "name": "importer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Importer Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImporterOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin",
          "importers"
        ],
        "summary": "Update Route",
        "operationId": "update_route_admin_importers__importer_id__patch",
        "parameters": [
          {
            "name": "importer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Importer Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImporterUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImporterOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin",
          "importers"
        ],
        "summary": "Delete Route",
        "operationId": "delete_route_admin_importers__importer_id__delete",
        "parameters": [
          {
            "name": "importer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Importer Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/importers/{importer_id}/run": {
      "post": {
        "tags": [
          "admin",
          "importers"
        ],
        "summary": "Run Route",
        "description": "Kick off a run in the background. Returns immediately with the\nimporter id; poll `/{id}/runs` for the actual run row's progress.\n\nWe instantiate the importer up-front (without running it) so a\nmissing optional-dep package fails the request with a 422 +\npip-install hint instead of letting the background task swallow\nthe error into the run row.",
        "operationId": "run_route_admin_importers__importer_id__run_post",
        "parameters": [
          {
            "name": "importer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Importer Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunQueuedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/importers/{importer_id}/runs": {
      "get": {
        "tags": [
          "admin",
          "importers"
        ],
        "summary": "Runs Route",
        "operationId": "runs_route_admin_importers__importer_id__runs_get",
        "parameters": [
          {
            "name": "importer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Importer Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunOut"
                  },
                  "title": "Response Runs Route Admin Importers  Importer Id  Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/importers/types": {
      "get": {
        "tags": [
          "v1",
          "importers"
        ],
        "summary": "List Types",
        "description": "All importer types the registry knows about.",
        "operationId": "list_types_v1_importers_types_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImporterTypeOut"
                  },
                  "title": "Response List Types V1 Importers Types Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/importers/discover": {
      "post": {
        "tags": [
          "v1",
          "importers"
        ],
        "summary": "Discover Route",
        "description": "Probe the source with provided creds and list available items.\n\nNo DB writes \u2014 secrets stay in memory only. The wizard sends them\nagain on save (where they get encrypted).",
        "operationId": "discover_route_v1_importers_discover_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoverIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiscoverIn": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "secrets": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Secrets"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "DiscoverIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Proposal": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "add",
              "correct",
              "deprecate"
            ],
            "title": "Type"
          },
          "target_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Node Id",
            "description": "For 'correct'/'deprecate' \u2014 node being changed."
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content",
            "description": "The proposed fact or correction, in natural language."
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence",
            "description": "Sources backing the claim \u2014 URLs, logs, version refs."
          },
          "rationale": {
            "type": "string",
            "title": "Rationale",
            "default": ""
          },
          "submitted_by": {
            "type": "string",
            "title": "Submitted By",
            "description": "Free-form attribution tag \u2014 stored verbatim, never trusted.",
            "default": "anonymous"
          }
        },
        "type": "object",
        "required": [
          "type",
          "content"
        ],
        "title": "Proposal"
      },
      "ProposalReceipt": {
        "properties": {
          "proposal_id": {
            "type": "string",
            "title": "Proposal Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "evaluating",
              "accepted",
              "rejected",
              "escalated",
              "errored"
            ],
            "title": "Status",
            "default": "queued"
          }
        },
        "type": "object",
        "required": [
          "proposal_id"
        ],
        "title": "ProposalReceipt"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Lighthouse-Key"
      }
    }
  }
}